+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### Latest
 
+- Cleanup of preprocessing config [!609](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/609)
 - Update puma version to v0.1.7 [!614](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/614)
 
 ### [v0.11](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/tags/0.11) (10.08.2022)